added some back styles
added back styles but didnt actually add a way to see it. Might add hover to flip animation at some point. Might make it so the print will drop the backs of the cards on the next page so you can print these double sided... might forget I did any of this in the morning and do something else instead. Who knows.